+++ apolib_controller.py
@@ -1,4 +1,4 @@
-__version__ = (0, 1, 0)
+__version__ = (0, 1, 2)
 
 
 # ▄▀█ █▄ █ █▀█ █▄ █ █▀█ ▀▀█ █▀█ █ █ █▀
@@ -17,8 +17,10 @@ __version__ = (0, 1, 0)
 
 # scope: hikka_only
 # scope: hikka_min 1.2.11
+# requires: emoji
 
 import logging
+import emoji  # skipcq: PY-W2000
 
 from telethon.tl.types import Message