+++ langreplier.py
@@ -1,4 +1,4 @@
-__version__ = (0, 1, 15)
+__version__ = (0, 1, 16)
 
 
 # ▄▀█ █▄ █ █▀█ █▄ █ █▀█ ▀▀█ █▀█ █ █ █▀
@@ -17,12 +17,13 @@ __version__ = (0, 1, 15)
 
 # scope: hikka_only
 # scope: hikka_min 1.2.11
-# requires: alphabet-detector
+# requires: alphabet-detector emoji googletrans==4.0.0-rc1
 
 import asyncio
 import logging
 import time
 
+import emoji  # skipcq: PY-W2000
 import googletrans
 from alphabet_detector import AlphabetDetector
 from telethon.tl.types import Message